requirements.txt
scikit-learn==1.1.1
pyyaml==6.0
xgboost==1.6.1
pandas==1.4.3
openpyxl==3.0.10